Split idl.rs into submodules
